backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.1
django-environ>=0.11
django-auth-ldap>=4.6
redis>=5.0
celery>=5.3
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.1
